# openapi.yaml
openapi: 3.0.3
info:
  title: Decibel Tribe API
  description: API specification for the Decibel Tribe application.
  version: 0.1.0
servers:
  - url: /api
    description: Local development server
paths:
...
  /posts/{postId}:
    parameters:
      - name: postId
        in: path
        required: true
        description: The ID of the post to operate on.
        schema:
          type: string
    get:
      summary: Get a specific post by ID
      description: Retrieves the details of a single post. Allows anonymous access but includes user-specific details (likes, dislikes, bookmarks) if authenticated.
      tags:
        - Posts
      security: [] # Allows anonymous access
      responses:
        "200":
          description: The requested post details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostData"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"
    patch:
      summary: Update a post
      description: Updates the content of an existing post. Requires authentication and user must be the owner of the post.
      tags:
        - Posts
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
                  description: The new content for the post.
              required:
                - content
      responses:
        "200":
          description: The updated post details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostData"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      summary: Delete a post
      description: Deletes an existing post. Requires authentication and user must be the owner of the post.
      tags:
        - Posts
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Post successfully deleted. Returns the deleted post data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostData" # Code returns deleted post
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /events/{eventId}:
    parameters:
      - name: eventId
        in: path
        required: true
        description: The ID of the event.
        schema:
          type: string
          format: uuid # Assuming event IDs are UUIDs
    get:
      summary: Get a specific event by ID
      description: Retrieves the details of a single event. Requires authentication. Returns 403 if the event is a draft and the user is not the owner. Includes attendee count and basic creator info.
      tags:
        - Events
      security:
        - cookieAuth: []
      responses:
        "200":
          description: The requested event details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventData"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          description: User does not have permission to view this draft event.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: Unauthorized # Actual error message from code
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"
    patch:
      summary: Update event details or Attend/Unattend
      description: >
        Updates the details of an existing event OR adds/removes the authenticated user as an attendee.
        Requires authentication. User must be the owner to update event details.
        To update details, provide the fields to update in the request body.
        To attend/unattend, provide `{"action": "attend"}` or `{"action": "unattend"}` in the request body.
      tags:
        - Events
        - Attendees
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: "#/components/schemas/UpdateEventInput" # Schema for updating details (fields optional)
                - $ref: "#/components/schemas/EventActionInput" # Schema for attend/unattend action
      responses:
        "200":
          description: Event updated or Attend/Unattend action successful. Returns the updated event data or attendee record.
          content:
            application/json:
              schema:
                oneOf: # Response depends on action
                  - $ref: "#/components/schemas/EventData" # If update
                  - $ref: "#/components/schemas/EventAttendee" # If attend/unattend
        "400":
          description: Invalid input data or invalid action.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                invalidData: { value: { error: "Invalid data", details: [] } }
                invalidAction: { value: { error: "Invalid action" } }
                alreadyAttending: { value: { error: "User is already an attendee of this event" } }
                notAttending: { value: { error: "User is not an attendee of this event" } }
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError" # User is not the owner for update
        "404":
          $ref: "#/components/responses/NotFoundError" # Event not found
        "500":
          $ref: "#/components/responses/InternalServerError"
    put:
      summary: Update event details (with cancellation notification)
      description: >
        Updates the details of a specific event by ID using PUT semantics (potentially full replacement, though current code behaves like PATCH).
        Requires authentication, and the user must be the owner of the event.
        If the `isCancelled` field changes from `false` to `true`,
        it sends `EVENT_CANCELLED` notifications to attendees (excluding the owner).
      tags:
        - Events
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEventInput" # Expects event fields to update
      responses:
        "200":
          description: Event updated successfully. Returns the updated event data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventData"
        "400":
          $ref: "#/components/responses/BadRequestError" # For invalid input
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError" # User is not the owner
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      summary: Delete an event
      description: Deletes a specific event. Requires authentication, and the user must be the owner of the event.
      tags:
        - Events
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Event deleted successfully. Returns the deleted event data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventData"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError" # User is not the owner
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: A message describing the error.
      required:
        - error
    UserSummary: # Basic user info used for embedding (based on getUserDataSelect)
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        displayName:
          type: string
          nullable: true
        avatarUrl:
          type: string
          format: url
          nullable: true
        createdAt: # Added from getUserDataSelect
          type: string
          format: date-time
        isAdmin: # Added from getUserDataSelect
          type: boolean
        _count: # Added from getUserDataSelect
          type: object
          properties:
            posts:
              type: integer
              description: Count of posts by this user.
            followers:
              type: integer
              description: Count of users following this user.
            following:
              type: integer
              description: Count of users this user is following.
          required: [posts, followers, following]
      required:
        - id
        - username
        - createdAt
        - isAdmin
        - _count
    UserWithFollowerStatus: # Full user profile data (extends UserSummary, adds isFollowedByMe)
      allOf:
        - $ref: "#/components/schemas/UserSummary"
        - type: object
          properties:
            isFollowedByMe:
              type: boolean
              description: Indicates if the logged-in user follows this user. Only present if logged in.
          # required: [isFollowedByMe] # Only present if logged in, so not strictly required
    Media: # Attachment type
      type: object
      properties:
        id:
          type: string
        url:
          type: string
          format: url
        type:
          type: string
          enum: [IMAGE, VIDEO] # Based on Prisma schema
        postId:
          type: string
          nullable: true # May not be associated yet during upload
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - url
        - type
        - createdAt
        - updatedAt

    # ---------- Post Related Schemas ----------
    PostData: # Represents a single post with included data (based on getPostDataInclude)
      type: object
      properties:
        id:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        userId:
          type: string
        groupId:
          type: string
          nullable: true
        user: # Embedded user data
          $ref: "#/components/schemas/UserWithFollowerStatus"
        attachments: # Array of media objects
          type: array
          items:
            $ref: "#/components/schemas/Media"
        Group: # Embedded group data (if post belongs to a group)
          type: object
          nullable: true
          properties:
            id:
              type: string
            name:
              type: string
          required: [id, name]
        _count: # Counts for post interactions
          type: object
          properties:
            likes:
              type: integer
            dislikes:
              type: integer
            comments:
              type: integer
          required: [likes, dislikes, comments]
        # Conditional fields (only included if user is logged in)
        likes:
          type: array
          description: Included if user is logged in; contains an entry if the user liked the post.
          items:
            type: object
            properties:
              userId:
                type: string
            required: [userId]
        dislikes:
          type: array
          description: Included if user is logged in; contains an entry if the user disliked the post.
          items:
            type: object
            properties:
              userId:
                type: string
            required: [userId]
        bookmarks:
          type: array
          description: Included if user is logged in; contains an entry if the user bookmarked the post.
          items:
            type: object
            properties:
              userId:
                type: string
            required: [userId]
      required:
        - id
        - content
        - createdAt
        - updatedAt
        - userId
        - user
        - attachments
        - _count
...
    EventAttendee: # Basic attendee record
      type: object
      properties:
        userId:
          type: string
        eventId:
          type: string
      required: [userId, eventId]
    EventAttendeeWithUser: # Attendee record including UserSummary
       allOf:
        - $ref: "#/components/schemas/EventAttendee"
        - type: object
          properties:
            user:
              $ref: "#/components/schemas/UserSummary"
          required: [user]
    EventData: # Represents event details (based on getEventDataInclude)
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        location:
          type: string
        url:
          type: string
          format: url
          nullable: true
        when:
          type: string
          format: date-time
        startTime:
          type: string
          pattern: '^([01]?[0-9]|2[0-3]):[0-5][0-9]$' # HH:mm format
          example: "14:30"
        endTime:
          type: string
          pattern: '^([01]?[0-9]|2[0-3]):[0-5][0-9]$' # HH:mm format
          example: "16:00"
        performers:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [DRAFT, PUBLISHED] # Based on Prisma schema/usage
        visibility:
          type: string
          enum: [PRIVATE, PUBLIC] # Based on Prisma schema/usage
        isCancelled:
          type: boolean
        createdById:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        createdBy: # Embedded user data (owner)
          $ref: "#/components/schemas/UserSummary" # Uses getUserDataSelect
        attendees: # List of attendees with user details
          type: array
          items:
            $ref: "#/components/schemas/EventAttendeeWithUser" # Uses getUserDataSelect for user
        _count: # Count of attendees
          type: object
          properties:
            attendees:
              type: integer
          required: [attendees]
      required:
        - id
        - title
        - location
        - when
        - startTime
        - endTime
        - performers
        - status
        - visibility
        - isCancelled
        - createdById
        - createdAt
        - updatedAt
        - createdBy
        - attendees
        - _count
...
    UpdateEventInput: # Schema for PATCH/PUT /events/{eventId} request body (update case)
       type: object
       properties:
         title:
           type: string
         location:
           type: string
         description:
           type: string
           nullable: true
         url:
           type: string
           format: url
           nullable: true
         when:
           type: string
           format: date-time
         startTime:
           type: string
           pattern: '^([01]?[0-9]|2[0-3]):[0-5][0-9]$' # HH:mm format
         endTime:
           type: string
           pattern: '^([01]?[0-9]|2[0-3]):[0-5][0-9]$' # HH:mm format
         performers:
           type: array
           items:
             type: string
         status:
           type: string
           enum: [DRAFT, PUBLISHED]
         visibility:
           type: string
           enum: [PRIVATE, PUBLIC]
         isCancelled: # Added for PUT handler, optional for PATCH
           type: boolean
       # No fields are strictly required for PATCH, any subset can be provided
       # PUT might imply all fields, but current code doesn't enforce
    EventActionInput: # Schema for PATCH /events/{eventId} request body (action case)
      type: object
      properties:
        action:
          type: string
          enum: [attend, unattend]
      required:
        - action
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_session # Use the actual session cookie name from lucia config
      description: "Authentication cookie obtained via login or OAuth callback. Send requests with `credentials: 'include'` on the client."
    cronAuth: # New scheme for cron job Bearer token
      type: http
      scheme: bearer
      bearerFormat: plain # Assuming the secret is used directly
      description: Bearer token authentication required for cron job endpoints (using CRON_SECRET environment variable value).

  # ---------- Standard Responses ----------
  responses:
    UnauthorizedError:
      description: Authentication information is missing or invalid (e.g., no session cookie, invalid session).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: Unauthorized
    ForbiddenError:
      description: Authenticated user does not have permission to perform the action (e.g., not owner, not admin, not group member).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: Forbidden
    NotFoundError:
      description: The requested resource (e.g., post, event, user, group) was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: Not Found # Message might be more specific like "Post not found"
    BadRequestError:
       description: The request is invalid (e.g., missing required parameters, invalid data format, validation error).
       content:
         application/json:
           schema:
             $ref: "#/components/schemas/Error" # Or a more detailed validation error schema
           example:
             error: Invalid input # Or validation details
    InternalServerError:
      description: An unexpected error occurred on the server.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: Internal server error